move shared objects to new structure
