+++ src/app/features/administration/assignments/assignments.component.ts
@@ -1,6 +1,6 @@
 import { Component, Input, OnInit, OnChanges } from 'angular2/core';
 import {AgGridNg2} from 'ag-grid-ng2/main';
-import {DataFetchingService} from '../../../shared/services/datafetching.service';
+import {DataFetchingService} from '../../../shared/features/services/datafetching.service';
 import {Store} from '../../../store/store';
 
 @Component({
